Optimized even odd algorthim
Optimized even odd algorithm to take advantage of python numeric singleton behavior.  All numbers <256 are references to the same underlying objects.  So `is` wll be faster for the check as we only have to compare the pointers and not call to `__eq__`.

Let me know when you publish so I can publish my optimization!